+++ src/test/java/com/google/code/guice/repository/testing/junit/general/CustomerRepositoryTest.java
@@ -40,7 +40,7 @@ public class CustomerRepositoryTest extends RepoTestBase {
     @Test
     public void testRepo() throws Exception {
         customerRepository.save(new Customer("name", "surname"));
-        customerRepository.sharedCustomMethod((long) 42);
+        customerRepository.sharedCustomMethod(42L);
 
         assertEquals("Invalid repository size", 1, customerRepository.count());
         assertEquals("Invalid repository size", 1, customerRepository.findAll(new PageRequest(0, 10)).getContent().size());
